Move html template generation to separate function
